refactor(fluid): move logic from math.go into pool_simulator.go
